refactor: move status to users
