+++ backend/src/app.module.ts
@@ -13,7 +13,7 @@ import { TypeOrmConfigModule } from './config/typeorm-config.module';
 import { FriendRequestModule } from './friend_requests/friend_request.module';
 import { GuestCleanupService } from './guest-cleanup/guest-cleanup.service';
 import { ScheduleModule } from '@nestjs/schedule';
-import { StatusModule } from './status/status.module';
+import { StatusModule } from './users/status/status.module';
 
 @Module({
   imports: [
