+++ backend/src/users/status/status.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { StatusGateway } from './status.gateway';
-import { UsersService } from '../users/users.service';
-import { SessionsService } from '../sessions/sessions.service';
-import { Session, User } from '../core/entities';
+import { UsersService } from '../users.service';
+import { SessionsService } from '../../sessions/sessions.service';
+import { Session, User } from '../../core/entities';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { S3ClientProvider } from '../lib/aws/s3Client';
+import { S3ClientProvider } from '../../lib/aws/s3Client';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User, Session])],
